requirements.txt
coverage==4.0.3
coveralls==1.1
dj-database-url==0.4.1
Django==1.9.5
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-generic-positions==0.2.2
django-hvad==1.5.0
django-libs==1.67.4
django-rest-swagger==0.3.6
django-user-media==1.2.2
djangorestframework==3.3.3
docopt==0.6.2
easy-thumbnails==2.3
gunicorn==19.4.5
Pillow==3.2.0
psycopg2==2.6.1
PyYAML==3.11
requests==2.9.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.19
stripe==1.32.2
whitenoise==3.0
